backend-python/requirements.txt
# Python Backend Dependencies
# Phase 0-4: Multi-Backend LLM + RAG + Tracing

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# HTTP Client (for LLM calls)
httpx>=0.26.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0

# Authentication (Phase 1)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Email (Phase 1)
aiosmtplib>=3.0.0

# Validation
email-validator>=2.1.0

# ======================================
# Phase 4: RAG + Tracing
# ======================================

# Vector embeddings
sentence-transformers>=2.2.0

# Vector databases
faiss-cpu>=1.7.4
pgvector>=0.2.0

# Redis for working memory
redis>=5.0.0

# Document processing
python-multipart>=0.0.6
pypdf>=4.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# ======================================
# Phase 6: MCP Integration
# ======================================
mcp>=1.0.0

# ======================================
# Phase 8: Agents Advanced
# ======================================
duckduckgo-search>=6.0.0

# ======================================
# Phase 9: Security
# ======================================
slowapi>=0.1.9
PyJWT>=2.7.0

# ======================================
# Phase 10: S3-Compatible Storage
# ======================================
boto3>=1.34.0

# ======================================
# Phase A: Enterprise Security
# ======================================
hvac>=1.2.0  # HashiCorp Vault client

# ======================================
# Phase B: Observability (OTel)
# ======================================
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-otlp>=1.20.0

# ======================================
# Phase E: Evaluation Framework
# ======================================
deepeval>=0.21.0
sentence-transformers>=2.2.0  # Embedding-based metrics
ragas>=0.1.0  # RAG evaluation (optional)
arize-phoenix>=3.0.0  # LLM observability (optional)
